Change `CFBundleIdentifier` to “net.culater.SIMBL” on `EasySIMBL.osax`
#12
Google Chrome has a "Bundle Blocker" mechanism which prevent loading unknown bundles.
The Bundle Blocker also have white list to allowing loading bundles.
Because the original SIMBL is in the white list, Google Chrome loads it.
http://src.chromium.org/viewvc/chrome/branches/874/src/chrome/common/mac/cfbundle_blocker.mm?r1=101122&r2=105568
But, EasySIMBL is not in the white list, so Google Chrome does not load it.